+++ framework/db-core/src/main/java/is/codion/framework/db/package-info.java
@@ -23,4 +23,5 @@
  * <li>{@link is.codion.framework.db.EntityConnectionProvider#CLIENT_CONNECTION_TYPE}
  * </ul>
  */
+@org.jspecify.annotations.NullMarked
 package is.codion.framework.db;